REGRESSION (r189567): Inline controls on Mac are misplaced https://bugs.webkit.org/show_bug.cgi?id=151072
Reviewed by Eric Carlson.

Source/WebCore:

Test: media/controls/fullscreen-button-inline-layout.html

For a certain range of video widths, r189567 caused the timeline track to be
too large, pushing the buttons on the right side of the inline controls out of
the rect bounds.

The fix is to set min-width to 0. The test added checks to see that the right
edge of the fullscreen button is within the rect bounds.

* Modules/mediacontrols/mediaControlsApple.css:
(audio::-webkit-media-controls-timeline-container): Add min-width.
(audio::-webkit-media-controls-panel .thumbnail-track): Ditto. Fly-by fix of height rule.

LayoutTests:

* media/controls/controls-test-helpers.js:
(ControlsTest.prototype.isLessThan): Add helper function.
* media/controls/fullscreen-button-inline-layout-expected.txt: Added.
* media/controls/fullscreen-button-inline-layout.html: Added.


Identifier: 169282@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@192205 268f45cc-cd09-0410-ab3c-d52691b4dbfc