Remove an unnecessary zero-delay timer from swipe snapshot removal https://bugs.webkit.org/show_bug.cgi?id=151161
Reviewed by Anders Carlsson.

* WebProcess/WebPage/ViewGestureGeometryCollector.cpp:
(WebKit::ViewGestureGeometryCollector::mainFrameDidLayout):
(WebKit::ViewGestureGeometryCollector::ViewGestureGeometryCollector): Deleted.
(WebKit::ViewGestureGeometryCollector::renderTreeSizeNotificationTimerFired): Deleted.
* WebProcess/WebPage/ViewGestureGeometryCollector.h:
There's no discernable reason for this zero-delay timer, and it dates back
to the original swipe patch with no explanation given. My only theory
is that I was depending on spinning the runloop to ensure we've painted,
but we have other mechanisms to do that now, on both platforms.

This also was a surprising waste of time in some cases where the web
process was very slow.


Identifier: 169350@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@192332 268f45cc-cd09-0410-ab3c-d52691b4dbfc