Unreviewed. Use commit id instead of tag name for jhbuild revision.
In r192469 I updated the jhbuild version to 3.12.0, and used the
tag instead of the commit id to make it easier to read and
remember. But we use the commit id to check if the current
revision is up to date, and when compared to the tag id it always
fails and jhbuild is always updated.

* jhbuild/jhbuild-wrapper: Use commit id and add a comment to
remember this is the id of 3.12.0.

Identifier: 169470@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@192472 268f45cc-cd09-0410-ab3c-d52691b4dbfc