[WinCairo][MediaFoundation] The main thread can sometimes be blocked forever when ending a media session. https://bugs.webkit.org/show_bug.cgi?id=151387
Reviewed by Alex Christensen.

This happens because the main thread is waiting for the sample scheduler thread to finish,
but sometimes it never does. This can be resolved by emptying the scheduler's sample queue
before requesting the scheduler thread to finish. This will make sure the scheduler thread
returns to its main loop quickly, and can handle the termination request.

* platform/graphics/win/MediaPlayerPrivateMediaFoundation.cpp:
(WebCore::MediaPlayerPrivateMediaFoundation::VideoScheduler::stopScheduler):


Identifier: 169549@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@192578 268f45cc-cd09-0410-ab3c-d52691b4dbfc