Tatechuyoko in ruby sits too high https://bugs.webkit.org/show_bug.cgi?id=151309 <rdar://problem/23536621>
Reviewed by Darin Adler.

Source/WebCore:

When combining text, we ask what the text's width is in order to determine if it fits in the
column. However, when we do that, we were not setting the font's orientation to horizontal.
This means that, for CJK text, the "width" which was returned was actually the height of the
glyph, and the GlyphOverflow data was similarly garbled.

We actually already were creating a corrected FontDescription, and using it in two places.
However, we weren't using it in the last place, which was causing this bug.

Test: fast/text/text-combine-placement.html

* platform/graphics/FontCascade.cpp:
(WebCore::FontCascade::width):
* rendering/RenderCombineText.cpp:
(WebCore::RenderCombineText::combineText):

LayoutTests:

* fast/text/resources/tatechuyoko.svg: Added.
* fast/text/text-combine-placement-expected.html: Added.
* fast/text/text-combine-placement.html: Added.


Identifier: 169608@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@192639 268f45cc-cd09-0410-ab3c-d52691b4dbfc