js/regress/script-tests/ftl-object-sub.js needs more iterations for benchmarking. https://bugs.webkit.org/show_bug.cgi?id=151411
Reviewed by Filip Pizlo.

Previously, in r192392, I reduced the test iteration count to make the layout
test happy and not time out.  Unfortunately, this defeats the use of this test
for benchmarking.  I've changed it to now use different iteration counts
depending on who is running it.  run-jsc-stress-tests is fine with the higher
iteration count.

Since this test is slow and only meant to test the FTL, I also changed the run
configuration to be runFTLNoCJIT to minimize the slowness hit when running jsc
stress tests.

* js/regress/script-tests/ftl-object-sub.js:



Identifier: 169625@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@192663 268f45cc-cd09-0410-ab3c-d52691b4dbfc