[GTK] UI process crash when the screensaver DBus proxy is being created while the web view is destroyed https://bugs.webkit.org/show_bug.cgi?id=151653
Reviewed by Martin Robinson.

We correctly cancel the proxy creation, but when the async ready
callback is called, the view could be destroyed already. In that
case g_dbus_proxy_new_for_bus_finish() will return nullptr and
fail with cancelled error, but we are using the passed web view
without checking first if the creation failed or not.

* UIProcess/API/gtk/WebKitWebViewBase.cpp:
(screenSaverProxyCreatedCallback):

Identifier: 169753@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@192792 268f45cc-cd09-0410-ab3c-d52691b4dbfc